Rename sources to what the URL specifies
Sources must be downloaded using wget, which creates oddly named sources, v1.7.2.tar.gz, though the source matches what rpmbuild and Tito expect due to the Source0 URL, which is derived from copying the link address of the hyperlink in the forge webpage for the current release. When downloading using a browser, the package name is prepended to the downloaded tarball (rpi-imager-1.7.2.tar.gz).